+++ Runtime/View/FilterViewExtensions.cs
@@ -10,51 +10,51 @@ namespace Massive
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Include<T>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<Include<T>, None>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<Include<T>, None>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Include<T1, T2>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<Include<T1, T2>, None>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<Include<T1, T2>, None>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Include<T1, T2, T3>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<Include<T1, T2, T3>, None>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<Include<T1, T2, T3>, None>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Include<T1, T2, T3, TInclude>(this View view)
 			where TInclude : IIncludeSelector, new()
 		{
-			return new FilterView(view.World, view.World.Filter<Include<T1, T2, T3, TInclude>, None>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<Include<T1, T2, T3, TInclude>, None>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Exclude<T>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<None, Exclude<T>>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<None, Exclude<T>>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Exclude<T1, T2>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<None, Exclude<T1, T2>>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<None, Exclude<T1, T2>>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Exclude<T1, T2, T3>(this View view)
 		{
-			return new FilterView(view.World, view.World.Filter<None, Exclude<T1, T2, T3>>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<None, Exclude<T1, T2, T3>>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		public static FilterView Exclude<T1, T2, T3, TExclude>(this View view)
 			where TExclude : IExcludeSelector, new()
 		{
-			return new FilterView(view.World, view.World.Filter<None, Exclude<T1, T2, T3, TExclude>>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<None, Exclude<T1, T2, T3, TExclude>>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -67,7 +67,7 @@ namespace Massive
 		public static FilterView Filter<TInclude>(this View view)
 			where TInclude : IIncludeSelector, new()
 		{
-			return new FilterView(view.World, view.World.Filter<TInclude, None>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<TInclude, None>(), view.PackingWhenIterating);
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
@@ -75,7 +75,7 @@ namespace Massive
 			where TInclude : IIncludeSelector, new()
 			where TExclude : IExcludeSelector, new()
 		{
-			return new FilterView(view.World, view.World.Filter<TInclude, TExclude>(), view.PackingWhenIterating);
+			return new FilterView(view.World, view.World.GetFilter<TInclude, TExclude>(), view.PackingWhenIterating);
 		}
 	}
 }