+++ Samples/ThumbnailSample.cs
@@ -29,10 +29,11 @@
 			var deltaTime = 1f / 60f;
 
 			// Iterate using lightweight views.
+			// The world itself acts as a view too.
 			var view = world.View();
 
-			// Views will select only those entities that contain all the necessary components.
-			view.ForEach((int entityId, ref Position position, ref Velocity velocity) =>
+			// ForEach will select only those entities that contain all the necessary components.
+			world.ForEach((int entityId, ref Position position, ref Velocity velocity) =>
 			{
 				position.Y += velocity.Magnitude * deltaTime;
 
@@ -49,7 +50,7 @@
 			});
 
 			// Pass arguments to avoid boxing.
-			view.ForEach((world, deltaTime),
+			world.ForEach((world, deltaTime),
 				(ref Position position, ref Velocity velocity,
 					(World World, float DeltaTime) args) =>
 				{
@@ -58,8 +59,7 @@
 
 			// Filter entities right in place.
 			// You don't have to cache anything.
-			world.View()
-				.Filter<Include<Player>, Exclude<Velocity>>()
+			world.Filter<Include<Player>, Exclude<Velocity>>()
 				.ForEach((ref Position position) =>
 				{
 					// ...
@@ -67,20 +67,20 @@
 
 			// Iterate using foreach with data set.
 			var positions = world.DataSet<Position>();
-			foreach (var entityId in world.View().Include<Player, Position>())
+			foreach (var entityId in world.Include<Player, Position>())
 			{
 				ref Position position = ref positions.Get(entityId);
 				// ...
 			}
 
-			// Chain any amount of components in filters.
+			// Chain any number of components in filters.
 			var filter = world.Filter<
 				Include<int, string, bool, Include<short, byte, uint, Include<ushort>>>,
 				Exclude<long, char, float, Exclude<double>>>();
 
-			// Reuse filter variable to reduce code duplication.
-			world.View().Filter(filter).ForEach((ref int n, ref bool b) => { });
-			world.View().Filter(filter).ForEach((ref string str) => { });
+			// Reuse the same filter view to iterate over different components.
+			filter.ForEach((ref int n, ref bool b) => { });
+			filter.ForEach((ref string str) => { });
 		}
 	}
 }