+++ Tests/Performance/FilterView3Include3IgnorePerformanceTest.cs
@@ -24,7 +24,7 @@ namespace Massive.PerformanceTests
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
 		private static Filter GetTestFilter(World world)
 		{
-			return world.Filter<
+			return world.GetFilter<
 				Include<TestState64, TestState64_2, TestState64_3>,
 				Exclude<TestState64<byte, int, int>, TestState64<int, byte, int>, TestState64<int, int, byte>>>();
 		}